backend/requirements.txt
fastapi
uvicorn
networkx
pandas
openai
python-dotenv
